if we are using doubles, why the heck should we use the float version of sin and cos? use the doubles, quell the warning.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11467 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/RobotPlayer.cxx
@@ -73,13 +73,13 @@ void RobotPlayer::projectPosition(const Player *targ,const float t,float &x,floa
     double hisspeed = hypotf(hisvx, hisvy);
     double alfa = omega * t;
     double r = hisspeed / fabs(omega);
-    double dx = r * sinf(alfa);
-    double dy2 = r * (1 - cosf(alfa));
+    double dx = r * sin(alfa);
+    double dy2 = r * (1 - cos(alfa));
     double beta = atan2(dy2, dx) * (targ->getAngularVelocity() > 0 ? 1 : -1);
     double gamma = atan2(hisvy, hisvx);
     double rho = gamma+beta;
-    sx = hisspeed * t * cosf(rho);
-    sy = hisspeed * t * sinf(rho);
+    sx = hisspeed * t * cos(rho);
+    sy = hisspeed * t * sin(rho);
   }
   x=(float)hisx+(float)sx;
   y=(float)hisy+(float)sy;
